requirements.txt
requests
beautifulsoup4
anthropic
google-generativeai
openai
python-dotenv
textstat
